Implement the ConsensusRunner
The ConsensusRunner is in charge of fetch processing events
from the database by passing the events to the Augrim algorithm
and storing the resulting actions. The actions are then passed
to the ConsensusActionRunner for processing.

Includes a builder for creating the runner and a unit tests that
processes a Start event.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>