Update the TimerHandler and TimerHandlerFactory
The TimerHandlerFactory will create a ConsensusRunner and
pass a ScabbardStore to the handler.

The TimerHandler will check to see if a TwoPhaseCommit alarm
has expires. If it has an Alarm event will be added to the store.

The ConsensusRunner will then be called to handle any pending
events and actions.

To properly handle the updates in splinterd, the creation of the
timer filter collection has been broken out into its own
file.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>