Add Identified<T> struct
Add an `identified` module containing the Identified<T> struct, this
tructs will be used to replace the IdentifiedConsensusAction and
IdentifiedConsensusEvent enums. The struct has two fields: id and
record, the `id` field has type i64 this is an ID associated with the
item stored in the `record` field, the `record` field has a generic
type so that the struct can be used for multiple types.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>