Update update_consensus_context
Update the `update_consensus_context` operation to remove the epoch
check before performing the update. Because of the changes in the
previous commit which removes epoch from the primary key for the consensus
context table, the `update_consensus_context` operation will be used
to update the existing context when a new epoch starts. The epoch check
would cause this operation to fail in that instance.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>