Update the update_consensus_context test
Previously the `scabbard_store_update_context` test attempted to update
a context with an epoch that doesn't exist and expected the operation to
fail. Because `update_consensus_context` has been updated to be able to
update the epoch of a context this operation should no longer be
expected to fail (and it does not). So this portion of the test is
removed.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>