Remove epoch arg from add_consensus_action
Update the add_consensus_action operation to remove the epoch arg
and instead get the epoch value from the current context in the db.

Remove the epoch arg from the SaveActionsCommand because it is no
longer needed now that add_consensus_action doesn't take epoch as
an arg. Also remove the epoch arg from the
ConsensusStoreCommandFactory's new_save_actions_command function,
which creates a new SaveActionsCommand.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>