Remove epoch arg from update_consensus_action
Remove the epoch arg from the update_consensus_action method. Update
the operation to just use action_id and service_id to get and update
the action.

Remove the epoch field from the ExecuteActionCommand. This command
calls the update_consensus_action function which no longer takes
epoch as an arg.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>