Guard service code with stable feature guards
The service features do not need to be default and should
only be pulled if services are being written/used.

Two features are added. "service" which guards the service
traits and "runtime-service" which guards all runtime service
components. This allows services to only pull in the parts of
the service API they need to implement the traits.

These features are added to stable.

This commit also adds a guard on the reply module in network,
as this is only used by the service instance code.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>