Add store and operation to ConstraintViolationType
Previously, the ConstraintViolationType depended on the
source error to provide helpful information about the
violation. This resulted in errors that were too cryptic
to be helpful.

This commit adds from_source_with_violation_type_and_store
which adds the store name and operation that will be prefixed
in front of the source error:

Store operation: source error

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>