Update ScabbarStoreError to include store and operation
ScabbardStoreError::from_source_with_operation that will set the
store name and operation type for the constraint violation errors.
The diesel operations are updated to use this constructor instead of
the default From implementation for the diesel errors.

Note the From implementation is left as that is required for the
use of connection.transaction

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>