Update Consensus Runner to run one action at a time
Previously, all actions were being passed to the action
runner. Due to the ordering of sending messages and executing
the store command returned, the ordering of actions was being
messed up.

This commit updates the Consensus Runner to pass one action at a
time and execute the commands before moving on to the next one.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>