Remove epoch from event, start event and vote event
Remove the epoch column/field from the tables and models for
consensus_2pc_event, consensus_2pc_start_event, and
consensus_2pc_vote_event. Only deliver events have an associated epoch
because of the message it contains so consensus_2pc_deliver_event should
be the only event related table with an epoch column.

Previously the epoch was being set for vote and start events based on
the epoch in the current context which caused a bug.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>