Remove epoch from actions where not needed
Remove the epoch column/field from the tables and models for
consensus_2pc_action, consensus_2pc_notification_action, and
consensus_2pc_update_context_action_participant. Send message and
update context actions are the only actions that will have an
associated epoch. All other action related tables do not need to have
an epoch column.

The epoch field in consensus_2pc_update_context_action_participant was
an unnecessary duplicate as it already exists for the associated update
context action in consensus_2pc_update_context_action and did not need
to be repeated so it was removed.

Previously the epoch for actions was being set using the current
context which caused a bug in scabbard consensus. This commit changes
the procedure to use the epoch stored in the action.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>