Fetch Context in ConsensusRunner right before use
This ensure the correct context will be used. Before the
context would fetched at the beginning of the call but
the context could have been updated by the following actions.

This commit moves the fetch to right before passing to the algorithm
to ensure it is the most up to date context.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>