Add new_unchecked public fn to Assignment
Manually constructing a struct outside of its own implementation or
possibly a builder is error prone and an anti pattern as it requires
every instance to understand the workings of the struct.

This commit adds a `new_unchecked` fn to `Assignment`s so that the
conversion between the database representation of an `Assignement` and
an internal `Assignment` doesn't require manually constructing the
struct. As the name implies there are no checks applied and the
resulting `Assignment` can be strictly speaking invalid. This is the
same behaviour as the code it replaces dictates and is a database
limitation we already have to work around.

Signed-off-by: Caleb Hill <hill@bitwise.io>