Update SyncLifecycleInterface to handle prepared service in add_service
Before if something had gone wrong when trying to prepare a service,
on restart to fix the preparing code there would be no component to
move the service on to finalized.

Now, on restart if the service has been added, it will be checked to see if
if the service has been prepared and needs to be finalized.

This works because the AdminService attempts to readd all services
on restart due to legacy service creation.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>