Move rbac/store to top level
As part of the greater effort to move the rest_api module to the common
and actix_web_1 crates the rbac store logic doesn't belong in either of
those crates. This commit moves the whole store implementation to the
new `rbac/store` module in libsplinter and then updates all usages to
point to that location.

Signed-off-by: Caleb Hill <hill@bitwise.io>