Update PermissionMap with crate friendly generics
As the users and definition of the `PermissionMap` struct are moved into
the actix_web_1 and common crates the current approach to adding support
for non actix-web-1 types has proven non optimal. The replaced
implementation was generic over types that could be converted to the
internal `Method` enum, which works really well within a single crate
because you can just add a `impl From<actix_web::Method> for Method`
block and call it a day.

The issue arrises when the `Method` is not a local type because rust
doesn't allow impl blocks for exclusively foreign types. Which usually
means you have to wrap the foreign types like so
`LocalWrapperMethod::new(actix_web::Method)` which is functional but
adds cruft and pretty much useless types.

The changes in this commit updates `PermissionMap`, and
`RequestDefinition` to be generic over the `Method` type. This is
possible because `PermissionMap`s don't need to handle multiple `Method`
types at the same time and can just use the raw library types without
issue.

Signed-off-by: Caleb Hill <hill@bitwise.io>