Add TransactionalUpgradeStores
This change adds the trait TransactionalUpgradeStores which adds the
additional behaviour of beginning a transaction.  The transaction
function takes a closure that accepts a reference to an UpgradeStores
dynamic instance that exists within the context of the transaction.

This solves the problem of starting and ending a transaction differently
than the way that grid does it (as of hyperledger/grid@64a2d5b), in that
it does not make use of an undocumented diesel API.  These APIs will not
be available in a forthcoming release.

It also implements the TransactionalUpgradeStores trait for sqlite,
postgres and the -WithLmdb wrapper.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>