Update Timer to keep track of which service are active
The Timer should only start a new TimerHandler/thread for a service
if one is not already running. To do this the Timer will keep track
of active services. The job will notify the Timer of when the
service's thread has completed by use a DeferedSend.

The DeferedSend sends the complete message on drop, and thus is also
triggered if the thread panics allow for proper clean up.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>