Move OauthResourceProvider to its own file
This commit moves the `OauthResourceProvider` struct into its own file.
The goal of this change is to make the eventual movement of the
enclosing module to the common/actix-web-1 crates. Having all the
structs in individual files makes that change a rename instead of a
combination delete/move/new file change.

Signed-off-by: Caleb Hill <hill@bitwise.io>