Remove scabbard store operation context checks
Remove the checks in the scabbard store operations that check to see
if an associated context with the given service ID exists and replace
it with a check to see if a service with the given service ID exists
in the database.

In the next two commits foreign keys linking consensus actions and
events to contexts are removed and foreign keys linking consensus
actions and events to services are added so checks need to be updated
to check the correct tables.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>