Remove ReactorError use statement in mod.rs
The use statement imports a type ( ReactorError ) not
defined/constrained to the module and makes its origin less clear in the
submodules that reimport it from the mod.rs file. This commit removes
the use statement in mod.rs and updates the use statement in
reactor_shutdown_signaler.rs to use a fuller path.

Signed-off-by: Caleb Hill <hill@bitwise.io>