Guarantee that mesh test ready before proceeding
This change uses barriers to ensure that each of the test threads waits
until the other has reached the same point before proceeding.  The first
point is adding the connections to mesh.  The second is receiving the
message from mesh, prior to shutting down.

The first barrier prevents a race condition where the message is sent
before the connection is added to the listening mesh, and therefore not
signaled about content on the socket.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>