Add helper scripts for psql migration test
check_splinterd: wait for splinterd nodes to be available. fail if they are not
                 ready in 10 seconds.

check_states: compare scabbard state root hashes.

prepare_circuit: create a circuit configured with the command family contract.

workload: parse out circuit and service IDs and run a short workload against
          those endpoints.

For now, the script to compare scabbard state root hashes will be used only
in the psql migration test but in the future could be genericized and moved
to a different location.

Signed-off-by: Ryan Beck-Buysse <rbuysse@bitwise.io>