Implement Supervisor that works with the ConsensusLayer
This commit adds a Supervisor that will handle pending notifications
that have been received from the consensus layer. The supervisor
currently handles each notification in the same way that the
CommandNotifyObserver currently does.

When a supervisor notification is added to state the Supervisor
should be notified. This must happen after the commands have been
executed, as such the notifier will be added to the timer handler.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>