Add PagingBuilder struct to paging mod
This commit adds the PagingBuilder struct and implementation to the
libsplinter rest_api module.

The usages of get_response_paging_info in libsplinter are going to be
replaced with the PagingBuilder once they move to the actix-web-1 crate.
To keep those changes as close to pure renames as possible updating
these files in libsplinter is really advantageous as they don't need to
be updated in the same commit as the planned moves.

Signed-off-by: Caleb Hill <hill@bitwise.io>