Remove get_response_paging_info fn from paging mod
This commit removes the get_response_paging_info fn from the paging
module as all its uses have been replaced with the PagingBuilder. This
includes updating the tests to use the PagingBuilder struct.

I did manually test that the get_response_paging_info fn and
PagingBuilder produce the same values given the same inputs. Moreover
looking at the changes in this commit you can see the PagingBuilder gets
the same inputs as the get_response_paging_info fn did and the
assertions have not changed. Given that the tests pass, this is about as
good of a check that their functionality is the same as we can hope for.

Signed-off-by: Caleb Hill <hill@bitwise.io>