Remove expect call in scabbard rest_api
This commit removes a call to `Option::expect` which panics if called on
an `Option::None`. This should not be possible in regular operation as
actix will not route requests to the handler if its missing that part of
the route. Better safe than sorry, this adds real error handling.

Signed-off-by: Caleb Hill <hill@bitwise.io>