Make scabbard SERVICE_TYPE const public
The `SERVICE_TYPE` is used by the rest api parts of the scabbard
service. As that bit of the scabbard code moves to the rest_api crates
the type either needs to be dulicated or made public. Duplication could
lead to weird out of sync errors if one is updated but not the other,
and there aren't really any downsides to making it public.

Signed-off-by: Caleb Hill <hill@bitwise.io>