Add orchestrator endpoint provider
This commit adds an `OrchestratorEndpointFactory` and a corresponding
builder. The structures are used within the orchestrator and its various
building stages. It serves to manage the endpoint providers passed to
the `ServiceOrchestratorBuilder` without being part of that struct.

The idea is that encapsulating that bit of functionality will make
removing all of it in favor of an external to the builder struct much
easier.

Signed-off-by: Caleb Hill <hill@bitwise.io>