Fix ServiceRequestGuard trait impl
The code this commit replaces could generate essentially endless cloned
boxes and would overflow the stack. This change replaces the trait with
a type declaration of an Arc wrapping the relevant trait object. This
makes it so we don't actually need to duplicate the trait object
everywhere and can just operate on references to to same one.

Signed-off-by: Caleb Hill <hill@bitwise.io>