Update models and schema for bool type
Update the consensus store models and schema to use type bool for the
vote and vote_response columns that were updated in the previous two
commits.

This change requires the implementation of `ListEventsOperation` to
be split up into two separate postgres and sqlite implementations
because of the use of type bool in queries.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>