Add builder for ServiceOrchestratorRestProvider
This commit updates splinterd to use the builder with the
`ScabbardServiceEndpointProvider` both in the daemon and node modules.

This change helps decouple the `ServiceOrchestrator` from whatever rest
endpoints ultimately need to be exposed. In the code removed in this
commit the `ServiceOrchestrator` provides rest endpoint providers, which
are then converted to actix structs/closures. This change makes it so
the `ServiceOrchestratorRestResourceProviderBuilder` takes rest endpoint
providers and filters them by what services are available in the passed
in `ServiceOrchestrator`.

Now the `ServiceOrchestrator` is free to not know anything at all about
the rest api besides the service types of the services it is managing.

Signed-off-by: Caleb Hill <hill@bitwise.io>