Remove loop from ConsensusRunner::run
This change removes the loop from the ConsensusRunner::run function.
This loop causes race conditions between the timer loop and this loop.
Removing the loop in the run function reduces the race condition surface
between the two.

It does run actions twice - once before the event, and once after the
event has executed.  The former ensures that any outstanding actions are
executed before the next event, in the case of a restart, the latter
ensures that any context changes are set after execution.  This latter
execution is important for initial event processing, due to the
overloading of the alarm as an "init" event.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>