Check the LMDB files exist before upgrading scabbard state
Without this check, if splinter is use postgres state, the upgrade
command created new lmdb files that would break a restart of splinter.

This commit adds a check to the upgrade command so it will return
OK without any changes if no LMDB files exists.

This makes the command able to be run on postgres only state for
easier scripting.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>