Update tests that use update_consensus_event
The update_consensus_event method gets the action_id of the most recent
update context action and stores it in an event after it has been
executed. This commit updates tests that called update_consensus_event
and didn't have any update context actions to add an update context
action and update it to appear as if it had been executed so that
update_consensus_event can use the associated action_id.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>