Add subcommands from Transact to splinter CLI
The transact CLI was added to provide a workload tool for
testing scabbard. We are working on removing dependencies from
transact and the transact CLI will be removed as transact is intended
to only be library.

Therefore, the transact CLI commands are being moved to the splinter CLI
as is. This also allows user to only require 1 CLI, instead of multiple.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>