Add postgres migration to fix pkey
Add a postgres migration to fix the primary key on the
`consensus_2pc_update_context_action_participant` table. The table
in postgres originally only had action_id as the primary key however
this was a bug as there can be multiple processes in the table with the
same action_id so process needs to be included in the key. The sqlite
table does not need to be updated as it already includes process in the
primary key. This was only an issue in postgres.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>