Make OAuthClient constructor public
OAuth integration tests make heavy use of the OAuthClient struct as you
might expect. Those integration tests are moving to the rest_api crates
because they also use actix-web specific code. This presents a problem
in that there wasn't a way to construct a client outside of libsplinter.

This commit makes the OAuthClient constructor public so it can be used
in the actix_web_1 crate.

Signed-off-by: Caleb Hill <hill@bitwise.io>